Conflicts:
	EmptyGeneralTesting/IntegrationTesting/QueryEvaluatorToPKBTest.cpp
	EmptyGeneralTesting/SPA/PKB/Parent.cpp
	EmptyGeneralTesting/SPA/PKB/VarTable.cpp
	EmptyGeneralTesting/SPA/Parser/Tokenizer.cpp
	EmptyGeneralTesting/SPA/Program/TNode/ProgramTNode.cpp
	EmptyGeneralTesting/SPA/Program/TNode/StmtListTNode.cpp
	EmptyGeneralTesting/UnitTesting/UsesTest.cpp

+++ EmptyGeneralTesting/SPA/PKB/VarTable.cpp
@@ -36,7 +36,7 @@ int VarTable::GetIndexOf(string varName) {
 }
 
 string VarTable::GetVarName(int varIndex) {
-	if (varIndex >= 0 && varIndex < varNames.size()) {
+	if (varIndex >= 0 && (unsigned int)varIndex < varNames.size()) {
 		return varNames[varIndex];
 	} else {
 		throw IndexNotFoundException();
